requirements.txt
ccxt>=4.2.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy<2.0
requests>=2.31.0
python-dotenv>=1.0.0
